pyproject.toml
[project]
name = "test-selfee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "ipython>=8.37.0",
    "ruff>=0.12.4",
]
